requirements.txt
asgiref==3.9.1
Django==4.2.24
django-cors-headers==4.7.0
djangorestframework==3.16.1
sqlparse==0.5.3
